Summary
- Fixed output from BaseParser.get_diff(...)
- Added optional max_length parameter, which is the max number of items in a rule
- Fixed print(...) arguments order
- Added 'local/' to .gitignore
- Added function get_commit_count(...)
- Printing absolute support value

Since the output of BaseParser.get_diff(...) is used to form the input of Apriori.get_rules(), the
output of get_diff(...) was incorrect. `get_rules()` expects a 'list of lists', but get_diff(...)
returns a 'list of list of tuples' like this:

[
  ('[doc.go] main',),
  ('[mux.go] main', '[mux.go] init', '[mux.go] setVars', '[mux.go] setCurrentRoute', '[mux.go] mapFromPairs'),
  ('[regexp.go] newRouteRegexp', '[regexp.go] braceIndices')
]

We cannot group changes in the same file in a tuple. The get_rules(...) would interpret each
tuple in this list as being an item of the transaction, but as you can see the second item contains
in fact multiple items (that happened in the same file, but multiple transactions nonetheless).
The correct input of a single transaction for get_rules for this same example is this:

[
  '[doc.go] main',
  '[mux.go] main',
  '[mux.go] init',
  '[mux.go] setVars',
  '[mux.go] setCurrentRoute',
  '[mux.go] mapFromPairs',
  '[regexp.go] newRouteRegexp',
  '[regexp.go] braceIndices'
]